refactor(local-inference): consolidate catalog in @elizaos/shared
The Eliza-1 model catalog (`MODEL_CATALOG`, `findCatalogModel`, tier ids,
default-eligibility set, HuggingFace URL builders) lived in two
byte-identical copies under `packages/ui/src/services/local-inference/`
and `packages/app-core/src/services/local-inference/`. Move the canonical
source to `@elizaos/shared/local-inference` and replace both call sites
with one-line re-export shims so existing `./catalog` and
`@elizaos/ui/...catalog` imports continue to resolve.

Net effect:
- 1 new file: `packages/shared/src/local-inference/catalog.ts` (canonical)
- 2 files reduced from 318 lines to 22-line shims
- All 51 UI + 273 app-core local-inference tests still pass against the
  shared catalog (one pre-existing downloader-manifest failure on the
  parent branch is unrelated to this change).

Note on Finding 3 (broader duplication sweep): the original plan listed
11 identical local-inference modules to move. Verification reduced the
real shareable target to `catalog.ts` only. Seven of the remaining
"identical" files (`active-model`, `bundled-models`, `dflash-doctor`,
`external-scanner`, `readiness`, `registry`, `routing-policy`) have zero
external consumers in the UI layer beyond their own tests and internal
cross-imports — they appear to be dead UI code that warrants a separate
audit before being moved or deleted. `hardware.ts` has a real 8-line
diff between layers and is not identical. `paths.ts` was already shimmed
in PR A.
